+++ backend/src/controllers/noticeController.js
@@ -1,5 +1,5 @@
 const noticeService = require("../services/noticeService");
-const multer = require('multer');
+
 const noticeView = (req, res) => {
 
     noticeService.getNotice(res);
@@ -33,31 +33,5 @@ const noticeUser = (req, res) => {
     noticeService.getUser(req, res);
 }
 
-const noticeImage = multer.diskStorage({
-	// (2)
-	destination: (req, file, cb) => {
-		// (3)
 
-		cb(null, '../../../Front-End/front/public/images/post_notice');
-	},
-	filename: (req, file, cb) => {
-		// (4)
-		imageNames.push(file.originalname);
-		cb(null, file.originalname); // (5)
-	},
-});
-
-const upload = multer({
-	// (6)
-	noticeImage,
-	fileFilter: (req, file, cb) => {
-		if (['image/jpeg', 'image/jpg', 'image/png'].includes(file.mimetype)) {
-			cb(null, true);
-		} else cb(new Error('해당 파일의 형식을 지원하지 않습니다.'), false);
-	},
-	limits: {
-		fileSize: 1024 * 1024 * 5,
-	},
-});
-
-module.exports = {noticeView, noticeDetail, noticeDelete, noticeModify, noticePost, noticeUser, upload}
+module.exports = {noticeView, noticeDetail, noticeDelete, noticeModify, noticePost, noticeUser}
