+++ backend/src/routes/noticeRouter.js
@@ -1,6 +1,34 @@
 const express = require('express');
 const router = express.Router();
 const noticeController = require("../controllers/noticeController.js");
+const multer = require('multer');
+
+const noticeImage = multer.diskStorage({
+	// (2)
+	destination: (req, file, cb) => {
+		// (3)
+
+		cb(null, '../../../Front-End/front/public/images/post_notice');
+	},
+	filename: (req, file, cb) => {
+		// (4)
+		imageNames.push(file.originalname);
+		cb(null, file.originalname); // (5)
+	},
+});
+
+const upload = multer({
+	// (6)
+	noticeImage,
+	fileFilter: (req, file, cb) => {
+		if (['image/jpeg', 'image/jpg', 'image/png'].includes(file.mimetype)) {
+			cb(null, true);
+		} else cb(new Error('해당 파일의 형식을 지원하지 않습니다.'), false);
+	},
+	limits: {
+		fileSize: 1024 * 1024 * 5,
+	},
+});
 
 //조회
 router.get("/view", noticeController.noticeView);
@@ -24,6 +52,8 @@ router.post("/post", noticeController.noticePost);
 
 router.get("/viewuser", noticeController.noticeUser);
 
-router.get("/image", noticeController.upload);
+router.post("/image", upload.single('image'), (req, res) => {
+    res.status(200).send({hi:"hello"});
+});
 
 module.exports = router;
